# api/service/api/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: carl.a.smigielski@saic.com
    name: Carl Smigielski
  description: An API for managing evaluations of Security Technical Implementation
    Guide (STIG) assessments.
  title: STIG Management API
  version: 1.0.0
servers:
- url: http://localhost:54000/stig-manager/api
paths:
...
  /stigs:
    get:
      operationId: getSTIGs
      parameters:
      - description: Package identifier
        explode: true
        in: query
        name: packageId
        required: false
        schema:
          type: integer
        style: form
      - description: Asset identifier
        explode: true
        in: query
        name: assetId
        required: false
        schema:
          type: integer
        style: form
      - description: A string found anywhere in a STIG title
        explode: true
        in: query
        name: title
        required: false
        schema:
          type: string
        style: form
      - description: An operating system TOE
        explode: true
        in: query
        name: os
        required: false
        schema:
          enum:
          - windows
          - linux
          - macos
          - ios
          - android
          - other
          - none
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/STIG'
                type: array
          description: STIG response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      security:
      - oauth:
        - stig-manager:stig:read
      summary: Return a list of available STIGs
      tags:
      - STIG
      x-swagger-router-controller: STIG
    post:
      operationId: addSTIG
      requestBody:
        content:
          multipart/form-data:
            schema:
              properties:
                source:
                  description: An XCCDF file
                  format: binary
                  type: string
        required: true
        x-name: body
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/STIG'
          description: The added STIG
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      security:
      - oauth:
        - stig-manager:stig
      summary: Import a new STIG
      tags:
      - STIG
      x-codegen-request-body-name: body
      x-rbac:
      - admin
      x-swagger-router-controller: STIG
  /stigs/{benchmarkId}:
    delete:
      operationId: deleteStigById
      parameters:
      - description: A path parameter that indentifies a STIG
        explode: false
        in: path
        name: benchmarkId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/STIG'
          description: The deleted STIG
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      security:
      - oauth:
        - stig-manager:stig
      summary: Deletes a STIG (*** and all revisions ***)
      tags:
      - STIG
      x-rbac:
      - admin
      x-swagger-router-controller: STIG
    get:
      operationId: getStigById
      parameters:
      - description: A path parameter that indentifies a STIG
        explode: false
        in: path
        name: benchmarkId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/STIG'
          description: STIG response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      security:
      - oauth:
        - stig-manager:stig:read
      summary: Return properties of the specified STIG
      tags:
      - STIG
      x-swagger-router-controller: STIG
  /stigs/{benchmarkId}/revisions:
    get:
      operationId: getRevisionsByBenchmarkId
      parameters:
      - description: A path parameter that indentifies a STIG
        explode: false
        in: path
        name: benchmarkId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Revision'
                type: array
          description: Revision response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      security:
      - oauth:
        - stig-manager:stig:read
      summary: Return a list of revisions for the specified STIG
      tags:
      - STIG
      x-swagger-router-controller: STIG
  /stigs/{benchmarkId}/revisions/{revisionStr}:
    delete:
      operationId: deleteRevisionByString
      parameters:
      - description: A path parameter that indentifies a STIG
        explode: false
        in: path
        name: benchmarkId
        required: true
        schema:
          type: string
        style: simple
      - description: A path parameter that indentifies a STIG revision [ V{version_num}R{release_num}
          | 'latest' ]
        explode: false
        in: path
        name: revisionStr
        required: true
        schema:
          pattern: ^(V\d+R\d+(\.\d+)?|latest)$
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Revision'
          description: The deleted Revision
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      security:
      - oauth:
        - stig-manager:stig
      summary: Deletes the specified revision of a STIG
      tags:
      - STIG
      x-rbac:
      - admin
      x-swagger-router-controller: STIG
    get:
      operationId: getRevisionByString
      parameters:
      - description: A path parameter that indentifies a STIG
        explode: false
        in: path
        name: benchmarkId
        required: true
        schema:
          type: string
        style: simple
      - description: A path parameter that indentifies a STIG revision [ V{version_num}R{release_num}
          | 'latest' ]
        explode: false
        in: path
        name: revisionStr
        required: true
        schema:
          pattern: ^(V\d+R\d+(\.\d+)?|latest)$
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Revision'
          description: Revision response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      security:
      - oauth:
        - stig-manager:stig:read
      summary: Return metadata for the specified revision of a STIG
      tags:
      - STIG
      x-swagger-router-controller: STIG
  /stigs/{benchmarkId}/revisions/{revisionStr}/groups:
    get:
      description: Can optionally specify profile filters.
      operationId: getGroupsByRevision
      parameters:
      - description: A path parameter that indentifies a STIG
        explode: false
        in: path
        name: benchmarkId
        required: true
        schema:
          type: string
        style: simple
      - description: A path parameter that indentifies a STIG revision [ V{version_num}R{release_num}
          | 'latest' ]
        explode: false
        in: path
        name: revisionStr
        required: true
        schema:
          pattern: ^(V\d+R\d+(\.\d+)?|latest)$
          type: string
        style: simple
      - description: Filter by profile
        explode: false
        in: query
        name: profile
        required: false
        schema:
          items:
            enum:
            - MAC-1_Classified
            - MAC-1_Sensitive
            - MAC-1_Public
            - MAC-2_Classified
            - MAC-2_Sensitive
            - MAC-2_Public
            - MAC-3_Classified
            - MAC-3_Sensitive
            - MAC-3_Public
            type: string
          type: array
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupList'
          description: Group response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      security:
      - oauth:
        - stig-manager:stig:read
      summary: Return the list of groups for the specified revision of a STIG.
      tags:
      - STIG
      x-swagger-router-controller: STIG
  /stigs/{benchmarkId}/revisions/{revisionStr}/groups/{groupId}:
    get:
      description: None
      operationId: getGroupByRevision
      parameters:
      - description: A path parameter that indentifies a STIG
        explode: false
        in: path
        name: benchmarkId
        required: true
        schema:
          type: string
        style: simple
      - description: A path parameter that indentifies a STIG revision [ V{version_num}R{release_num}
          | 'latest' ]
        explode: false
        in: path
        name: revisionStr
        required: true
        schema:
          pattern: ^(V\d+R\d+(\.\d+)?|latest)$
          type: string
        style: simple
      - description: A path parameter that indentifies a Group
        explode: false
        in: path
        name: groupId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupObj'
          description: Group response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      security:
      - oauth:
        - stig-manager:stig:read
      summary: Return the rules, checks and fixes for a Group from a specified revision
        of a STIG.
      tags:
      - STIG
      x-swagger-router-controller: STIG
  /stigs/{benchmarkId}/revisions/{revisionStr}/rules:
    get:
      description: Can optionally specify profile filters.
      operationId: getRulesByRevision
      parameters:
      - description: A path parameter that indentifies a STIG
        explode: false
        in: path
        name: benchmarkId
        required: true
        schema:
          type: string
        style: simple
      - description: A path parameter that indentifies a STIG revision [ V{version_num}R{release_num}
          | 'latest' ]
        explode: false
        in: path
        name: revisionStr
        required: true
        schema:
          pattern: ^(V\d+R\d+(\.\d+)?|latest)$
          type: string
        style: simple
      - description: Filter by profile
        explode: false
        in: query
        name: profile
        required: false
        schema:
          items:
            enum:
            - MAC-1_Classified
            - MAC-1_Sensitive
            - MAC-1_Public
            - MAC-2_Classified
            - MAC-2_Sensitive
            - MAC-2_Public
            - MAC-3_Classified
            - MAC-3_Sensitive
            - MAC-3_Public
            type: string
          type: array
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Rule'
                type: array
          description: Rule response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      security:
      - oauth:
        - stig-manager:stig:read
      summary: Return rule data for the specified revision of a STIG.
      tags:
      - STIG
      x-swagger-router-controller: STIG
  /stigs/rules/{ruleId}:
    get:
      operationId: getRuleByRuleId
      parameters:
      - description: A path parameter that indentifies a Rule
        explode: false
        in: path
        name: ruleId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
          description: Rule response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      security:
      - oauth:
        - stig-manager:stig:read
      summary: Return the defintion and associated checks and fixes for the specified
        Rule
      tags:
      - STIG
      x-swagger-router-controller: STIG
  /stigs/{benchmarkId}/revisions/{revisionStr}/ccis:
    get:
      operationId: getCcisByRevision
      parameters:
      - description: A path parameter that indentifies a STIG
        explode: false
        in: path
        name: benchmarkId
        required: true
        schema:
          type: string
        style: simple
      - description: A path parameter that indentifies a STIG revision [ V{version_num}R{release_num}
          | 'latest' ]
        explode: false
        in: path
        name: revisionStr
        required: true
        schema:
          pattern: ^(V\d+R\d+(\.\d+)?|latest)$
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CciListItem'
                type: array
          description: CCI response
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: unexpected error
      security:
      - oauth:
        - stig-manager:stig:read
      summary: Return a list of CCIs from a STIG revision
      tags:
      - STIG
      x-swagger-router-controller: STIG
...
components:
...
  schemas:
    Error:
      properties:
        code:
          format: int32
          type: integer
        message:
          type: string
      required:
      - code
      - message
      type: object
    STIG:
      example:
        lastRevisionDate: lastRevisionDate
        title: title
        benchmarkId: benchmarkId
        lastRevisionStr: lastRevisionStr
      properties:
        benchmarkId:
          type: string
        title:
          type: string
        lastRevisionStr:
          type: string
        lastRevisionDate:
          type: string
      required:
      - benchmarkId
      type: object
    Revision:
      example:
        statusDate: statusDate
        revisionStr: revisionStr
        release: release
        description: description
        benchmarkDate: benchmarkDate
        version: version
        benchmarkId: benchmarkId
        status: status
      properties:
        benchmarkId:
          type: string
        revisionStr:
          type: string
        version:
          type: string
        release:
          type: string
        benchmarkDate:
          type: string
        status:
          type: string
        statusDate:
          type: string
        description:
          type: string
      type: object
    GroupList:
      items:
        properties:
          groupId:
            type: string
          title:
            type: string
          profiles:
            items:
              $ref: '#/components/schemas/Profile'
            type: array
          rules:
            items:
              $ref: '#/components/schemas/RuleAbbr'
            type: array
        type: object
      type: array
    GroupObj:
      example:
        groupId: groupId
        profiles:
        - null
        - null
        rules:
        - severity: severity
          vulnDiscussion: vulnDiscussion
          thirdPartyTools: thirdPartyTools
          falseNegatives: falseNegatives
          documentable: documentable
          cci:
          - cci
          - cci
          weight: weight
          falsePositives: falsePositives
          potentialImpacts: potentialImpacts
          title: title
          version: version
          mitigations: mitigations
          fixes:
          - text: text
            fixId: fixId
          - text: text
            fixId: fixId
          mitigationControl: mitigationControl
          checks:
          - checkId: checkId
            content: content
          - checkId: checkId
            content: content
          responsibility: responsibility
          securityOverrideGuidance: securityOverrideGuidance
          ruleId: ruleId
        - severity: severity
          vulnDiscussion: vulnDiscussion
          thirdPartyTools: thirdPartyTools
          falseNegatives: falseNegatives
          documentable: documentable
          cci:
          - cci
          - cci
          weight: weight
          falsePositives: falsePositives
          potentialImpacts: potentialImpacts
          title: title
          version: version
          mitigations: mitigations
          fixes:
          - text: text
            fixId: fixId
          - text: text
            fixId: fixId
          mitigationControl: mitigationControl
          checks:
          - checkId: checkId
            content: content
          - checkId: checkId
            content: content
          responsibility: responsibility
          securityOverrideGuidance: securityOverrideGuidance
          ruleId: ruleId
        title: title
      properties:
        groupId:
          type: string
        title:
          type: string
        profiles:
          items:
            $ref: '#/components/schemas/Profile'
          type: array
        rules:
          items:
            $ref: '#/components/schemas/Rule'
          type: array
      type: object
    Profile:
      enum:
      - MAC-1_Classified
      - MAC-1_Sensitive
      - MAC-1_Public
      - MAC-2_Classified
      - MAC-2_Sensitive
      - MAC-2_Public
      - MAC-3_Classified
      - MAC-3_Sensitive
      - MAC-3_Public
      type: string
    Rule:
      example:
        severity: severity
        vulnDiscussion: vulnDiscussion
        thirdPartyTools: thirdPartyTools
        falseNegatives: falseNegatives
        documentable: documentable
        cci:
        - cci
        - cci
        weight: weight
        falsePositives: falsePositives
        potentialImpacts: potentialImpacts
        title: title
        version: version
        mitigations: mitigations
        fixes:
        - text: text
          fixId: fixId
        - text: text
          fixId: fixId
        mitigationControl: mitigationControl
        checks:
        - checkId: checkId
          content: content
        - checkId: checkId
          content: content
        responsibility: responsibility
        securityOverrideGuidance: securityOverrideGuidance
        ruleId: ruleId
      properties:
        ruleId:
          type: string
        version:
          type: string
        title:
          type: string
        severity:
          type: string
        weight:
          type: string
        vulnDiscussion:
          type: string
        falsePositives:
          type: string
        falseNegatives:
          type: string
        documentable:
          type: string
        mitigations:
          type: string
        securityOverrideGuidance:
          type: string
        potentialImpacts:
          type: string
        thirdPartyTools:
          type: string
        mitigationControl:
          type: string
        responsibility:
          type: string
        cci:
          items:
            type: string
          type: array
        checks:
          items:
            $ref: '#/components/schemas/Check'
          type: array
        fixes:
          items:
            $ref: '#/components/schemas/Fix'
          type: array
      type: object
    RuleAbbr:
      properties:
        ruleId:
          type: string
        version:
          type: string
        title:
          type: string
        severity:
          type: string
      type: object
    Check:
      example:
        checkId: checkId
        content: content
      properties:
        checkId:
          type: string
        content:
          type: string
      type: object
    Fix:
      example:
        text: text
        fixId: fixId
      properties:
        fixId:
          type: string
        text:
          type: string
      type: object
    CciReferenceItem:
      example:
        parentControl: parentControl
        creator: creator
        indexDisa: indexDisa
        textRefNist: textRefNist
        location: location
        title: title
        version: version
      properties:
        creator:
          type: string
        title:
          type: string
        version:
          type: string
        location:
          type: string
        indexDisa:
          type: string
        textRefNist:
          type: string
        parentControl:
          type: string
      type: object
    CciListItem:
      example:
        references:
        - parentControl: parentControl
          creator: creator
          indexDisa: indexDisa
          textRefNist: textRefNist
          location: location
          title: title
          version: version
        - parentControl: parentControl
          creator: creator
          indexDisa: indexDisa
          textRefNist: textRefNist
          location: location
          title: title
          version: version
        cci: cci
        type: type
      properties:
        cci:
          type: string
        type:
          type: string
        references:
          items:
            $ref: '#/components/schemas/CciReferenceItem'
          type: array
      type: object
...
  securitySchemes:
    oauth:
      flows:
        implicit:
          authorizationUrl: http://localhost:8080/auth/realms/cmsat/protocol/openid-connect/auth
          scopes:
            stig-manager: Top-level scope
            stig-manager:stig: Full access to the STIG collection
            stig-manager:stig:read: Read access to the STIG collection
            stig-manager:package: Full access to the Package collection
            stig-manager:package:read: Read-only access to the Package collection
            stig-manager:user: Full access to the Users collection
            stig-manager:user:read: Read-only access to the Users collection
      type: oauth2